pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kuramoto"
version = "0.2.0"
description = "Kuramoto oscillator simulation for 2D cortical sheets"
requires-python = ">=3.10"
dependencies = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "diffrax>=0.5.0",
    "equinox>=0.11.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]
readme = "README.md"

[tool.hatch.build.targets.wheel]
packages = ["kuramoto"]
